feat: reorganized components, added missing data
reorganized components, added decision-item component
