+++ frontend/src/app/constants/case_actions.ts
@@ -3,6 +3,7 @@ export const CASE_ACTION_TYPE = {
   PREVENTION_EDUCATION: "COSPRV&EDU",
   EQUIPMENT: "EQUIPMENT",
   WILDLIFE: "WILDLIFE",
+  CEEBACTION: "CEEBACTION",
 };
 
 export const CASE_ACTION_CODE = {
@@ -23,4 +24,8 @@ export const CASE_ACTION_CODE = {
   UPDATENOTE: "UPDATENOTE",
   ADMNSTRDRG: "ADMNSTRDRG",
   RECOUTCOME: "RECOUTCOME",
+  FWDLEADAGN: "FWDLEADAGN",
+  RESPNFA: "RESPNFA",
+  RESPAUTO: "RESPAUTO",
+  RESPREC: "RESPREC",
 };
