+++ frontend/src/app/types/app/case-files/ceeb/decision/decision-action-input.ts
@@ -1,0 +1,6 @@
+export interface DecisionActionInput {
+  id?: string;
+  actor: string;
+  action: string;
+  date?: Date;
+}
