+++ frontend/src/app/types/app/case-files/ceeb/decision/decision.ts
@@ -1,9 +1,12 @@
 export interface Decision {
+  id?: string;
   schedule: string;
   sector: string;
   discharge: string;
   nonCompliance: string;
   rationale: string;
+  inspectionNumber?: string;
+  leadAgency?: string;
   assignedTo: string;
   actionTaken: string;
   actionTakenDate: Date | null;
